Update patch set 2
Patch Set 2:

> I'll prefer not using macro to avoid modifying delorean to
 > understand the macro. That's only this package that is causing a
 > loop so I don't feel the energy to do such a big change in delorean
 > for only one package.

I'm afraid we may find the same situation later with another package, so we are just postponing the pain.

I'm going to try a patch I have in mind for Delorean (add a variable to the mock command-line when running with --order). If it works fine, it might be simpler for everyone.

Patch-set: 2
Reviewer: Javier Peña <1003941@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0